+++ Assets/Live2D/Cubism/Samples/OriginalWorkflow/Demo/CubismLookTarget.cs
@@ -8,6 +8,7 @@
 
 using Live2D.Cubism.Framework.LookAt;
 using UnityEngine;
+using UnityEngine.InputSystem;
 
 namespace Live2D.Cubism.Samples.OriginalWorkflow.Demo
 {
@@ -19,7 +20,7 @@ namespace Live2D.Cubism.Samples.OriginalWorkflow.Demo
         /// <returns>Mouse coordinates.</returns>
         public Vector3 GetPosition()
         {
-            var targetPosition = Input.mousePosition;
+            var targetPosition = (Pointer.current != null) ? (Vector3)Pointer.current.position.ReadValue() : Vector3.zero;
 
             var z = Camera.main.WorldToScreenPoint(transform.position).z;
             targetPosition.z = z;
@@ -34,7 +35,7 @@ namespace Live2D.Cubism.Samples.OriginalWorkflow.Demo
         /// <returns><see langword="true"/> if the target is active; <see langword="false"/> otherwise.</returns>
         public bool IsActive()
         {
-            return Input.GetMouseButton(0);
+            return Pointer.current != null && Pointer.current.press.isPressed;
         }
     }
 }