+++ Assets/Live2D/Cubism/Plugins/CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
+## 2026-04-02
+
+### Removed
+
+* [Unity] Remove Android x86 library.
+
+
+## 2026-01-29
+
+### Added
+
+* Add Arm64 static library for iphonesimulator.
+
+
 ## 2026-01-08
 
 ### Changed